esociety/requirements.txt
db==0.1.1
db-sqlite3==0.0.1
Django==3.1.3
django-bootstrap4==2.3.1
mysqlclient==2.0.1 
PyMySQL==0.9.3